+++ spec/rubocop/cop/rspec/match_array_spec.rb
@@ -23,6 +23,23 @@ RSpec.describe RuboCop::Cop::RSpec::MatchArray do
     RUBY
   end
 
+  it 'flags `match_array` with an empty array literal argument' do
+    expect_offense(<<-RUBY)
+      it { is_expected.to match_array([]) }
+                          ^^^^^^^^^^^^^^^ Prefer `eq` when matching an empty array literal.
+      it { is_expected.to match_array(%w()) }
+                          ^^^^^^^^^^^^^^^^^ Prefer `eq` when matching an empty array literal.
+      it { is_expected.to match_array(%i()) }
+                          ^^^^^^^^^^^^^^^^^ Prefer `eq` when matching an empty array literal.
+    RUBY
+
+    expect_correction(<<-RUBY)
+      it { is_expected.to eq([]) }
+      it { is_expected.to eq([]) }
+      it { is_expected.to eq([]) }
+    RUBY
+  end
+
   it 'does not flag `contain_exactly`' do
     expect_no_offenses(<<-RUBY)
       it { is_expected.to contain_exactly(content1, content2) }